==> 01-Clean-Episodes-SNF.ipynb <==
import numpy as np
import pandas as pd

SNFs = pd.read_excel('Data/SNFs.xlsx')

Episodes = pd.read_excel('Data/Episodes.xlsx')

SNFs.head()

Episodes.head()

# merge SNFs with Episodes and only keep SNF that has patient data in Episodes
# Reasons to do this: simplier question and, for the SNFs that do not have any patient record,
# no way to confirm that they need CCs and they need how many CCs.
common_epi_snf = SNFs.merge(Episodes, on=["PID"])
SNFs_no_episodes = SNFs[~SNFs.PID.isin(common_epi_snf.PID)]

common_snf = SNFs.merge(SNFs_no_episodes, on=["PID"])
SNFs_with_episodes = SNFs[~SNFs.PID.isin(common_snf.PID)]

# Align with unique address and city
repeat = SNFs_with_episodes[SNFs_with_episodes.duplicated(subset=['ProviderPracticeAddress1', 'ProviderPracticeCity'],keep=False)].sort_values(by ='ProviderPracticeAddress1')

common_both = SNFs_with_episodes.merge(repeat, on=["PID"])
SNFs_unique = SNFs_with_episodes[~SNFs_with_episodes.PID.isin(common_both.PID)]

repeat_for_join = repeat.groupby(['ProviderPracticeAddress1','County'], as_index = False).first()

frames = [SNFs_unique, repeat_for_join]
SNFs_fixed = pd.concat(frames, sort = False)

SNFs_fixed.to_csv('Data/SNFs.csv', index=False)

#Replacing wrong PIDs
Episodes['PID'] = Episodes['PID'].replace([560, 1770, 1632, 484, 147, 463, 804, 474, 837, 757, 135, 612, 16],[561, 1771, 1631, 485, 146, 464, 802, 478, 836, 38, 138, 613, 12])

# convert to datetime type
Episodes['AdmitDate']= pd.to_datetime(Episodes['AdmitDate'])
Episodes['DischargeDate']= pd.to_datetime(Episodes['DischargeDate'])

Episodes = Episodes.dropna(axis=0, subset=['DischargeDate'])

# have a new column store time difference
Episodes['time_diff'] = Episodes['DischargeDate'] - Episodes['AdmitDate']

Episodes = pd.DataFrame(Episodes)

Episodes = Episodes[Episodes['time_diff'].dt.days >= 0]

Episodes = Episodes.drop(['time_diff'], axis = 1)

Episodes.to_csv('Data/Episodes.csv', index=False)
