==> 02-get-episodes-duration.ipynb <==
import pandas as pd
import numpy as np

episodes = pd.read_csv("Data/Episodes.csv")

#change AdmitDate and DischargeDate to datetime so we can subtract one from the other
episodes['AdmitDate']= pd.to_datetime(episodes['AdmitDate']) 
episodes['DischargeDate']= pd.to_datetime(episodes['DischargeDate']) 

#Subtract DischargeDate from AdmitDate to get total number of days patient was in SNF
episodes['duration']= abs(episodes['AdmitDate']-episodes['DischargeDate'])
#change duration data type so we can manipulate it mathematically
episodes['duration']=episodes['duration'].astype('timedelta64[D]')

#filter columns
episodes_dur = episodes.filter(['PID', 'EpisodeID', 'ContractType', 'duration'])
episodes_dur.head()

#save to csv
episodes_dur.to_csv("Data/episodes_dur.csv")
