==> 03-Calculate-Num-CC.ipynb <==
import pandas as pd
import numpy as np

episodes_duration = pd.read_csv("Data/episodes_dur.csv")

episodes_duration.head()

episodes_duration = episodes_duration.drop('Unnamed: 0', axis = 1)

episodes_duration['num_mins'] = ""
episodes_duration.head()

for index, row in episodes_duration.iterrows():
    if row.ContractType == 'CT1':
        if row.duration <= 3:
            episodes_duration.loc[index, 'num_mins'] = 43
        elif row.duration > 3:
            Num_weeks = np.ceil(row.duration/7)
            episodes_duration.loc[index, 'num_mins'] = 31 + (Num_weeks * 34) + 12
    elif row.ContractType == 'CT2':
        if row.duration <= 3:
            episodes_duration.loc[index, 'num_mins'] = 117
        elif row.duration > 3:
            Num_weeks = np.ceil(row.duration/7)
            episodes_duration.loc[index, 'num_mins'] = 83 + (Num_weeks * 66) + 34

episodes_duration.head()

temp = episodes_duration.drop(['EpisodeID', 'ContractType', 'duration'], axis = 1)
temp.head()

temp = temp.groupby(['PID']).sum()
temp.head()

temp['num_hours'] = ''
temp['num_weeks'] = ''
temp['num_CC'] = ''

for index, row in temp.iterrows():
    temp.loc[index, 'num_hours'] = row.num_mins/60

for index, row in temp.iterrows():
    temp.loc[index, 'num_weeks'] = row.num_hours/52

for index, row in temp.iterrows():
    temp.loc[index, 'num_CC'] = row.num_weeks/40

temp

temp.to_csv("Data/CCs.csv", index = True)
