==> 04-Geocoding.ipynb <==
from googlemaps import Client as GoogleMaps
import pandas as pd 
import time

#PLEASE USE YOUR OWN GOOGLE MAPS API KEY HERE. 
#If you don't have one, the result of this ipynb is simply added columns for latitude and longitude for each SNF,
#saved in 'addresses.csv'
gmaps = GoogleMaps('INSERT KEY HERE')

addresses = pd.read_csv("Data/SNFs.csv")
addresses.head()

addresses['Full_Address'] = addresses['ProviderPracticeAddress1'].str.cat(addresses['ProviderPracticeCity'],sep=" ")
addresses.head()

addresses['Full_Address'] = addresses['Full_Address'].str.cat(addresses['ProviderPracticeStateCode'],sep=" ")
addresses.head()

addresses['ProviderPracticeZipCode'] = addresses['ProviderPracticeZipCode'].astype(str)

addresses['Full_Address'] = addresses['Full_Address'].str.cat(addresses['ProviderPracticeZipCode'],sep=" ")
addresses.head()

addresses['long'] = ""
addresses['lat'] = ""

addresses.head()

for x in range(len(addresses)):
    try:
        time.sleep(1) #to add delay in case of large DFs
        geocode_result = gmaps.geocode(addresses['Full_Address'][x])
        addresses['lat'][x] = geocode_result[0]['geometry']['location'] ['lat']
        addresses['long'][x] = geocode_result[0]['geometry']['location']['lng']
    except IndexError:
        print("Address was wrong...")
    except Exception as e:
        print("Unexpected error occurred.", e )
addresses.head()

addresses.to_csv('Data/addresses.csv', index = False)
