==> 05-get-average-duration-per-SNF-CT.ipynb <==
import pandas as pd
import numpy as np

episodes_dur = pd.read_csv('Data/episodes_dur.csv')

episodes_dur['num_mins'] = ""

for index, row in episodes_dur.iterrows():
    if row.ContractType == 'CT1':
        if row.duration <= 3:
            episodes_dur.loc[index, 'num_mins'] = 43
        elif row.duration > 3:
            Num_weeks = np.ceil(row.duration/7)
            episodes_dur.loc[index, 'num_mins'] = 31 + (Num_weeks * 34) + 12
    elif row.ContractType == 'CT2':
        if row.duration <= 3:
            episodes_dur.loc[index, 'num_mins'] = 117
        elif row.duration > 3:
            Num_weeks = np.ceil(row.duration/7)
            episodes_dur.loc[index, 'num_mins'] = 83 + (Num_weeks * 66) + 34

episodes_dur = pd.DataFrame(episodes_dur.drop('Unnamed: 0', axis = 1))

count_series = episodes_dur.groupby(['PID', 'ContractType']).size()

pat_num_PID_CT = count_series.to_frame(name = 'size').reset_index()

duration_PID_CT = episodes_dur.groupby(['PID', 'ContractType'], as_index=False)['num_mins'].sum()

Pat_duration_sum_SNF_CT = pat_num_PID_CT.merge(duration_PID_CT, on=['PID', 'ContractType'], how='inner')

Pat_duration_sum_SNF_CT["average_time_spent"] = Pat_duration_sum_SNF_CT["num_mins"]/Pat_duration_sum_SNF_CT["size"]

Pat_duration_sum_SNF_CT.to_csv('Data/avg_duration_per_SNF_CT.csv', index=False)
