==> 06-Clean-County-Data.ipynb <==
from datetime import datetime
import math
import numpy as np
import pandas as pd

SNF = pd.read_excel('Data/SNFs.xlsx')

zipcode = pd.read_csv('uszips.csv')

new = pd.merge(SNF,zipcode,how='left',left_on='ProviderPracticeZipCode', right_on='zip')
new.to_csv('Data/new_SNF.csv')

new['county_name'][1181] = 'Collin'

new['county_name'][1181]

new.to_csv("Data/county_complete.csv",index=False)
