==> 07-Get-Monthly-Patient-Data.ipynb <==
# load library
from datetime import datetime
import math
import numpy as np
import pandas as pd

# load data
csv_file = "Data/Episodes.csv"
csv_data = pd.read_csv(csv_file, low_memory = False)
csv_df = pd.DataFrame(csv_data)
Episodes_df1 = csv_df[["PID","EpisodeID"]]


#load data - Episodes_new is just the original excel file "Data/Episodes" changed to a csv file for better formatting.
csv_file = "Data/Episodes_new.csv"
csv_data = pd.read_csv(csv_file, low_memory = False)
csv_df = pd.DataFrame(csv_data)
Episodes_df2 = csv_df[["EpisodeID","ContractType","AdmitDate","DischargeDate"]]

Episodes_df = pd.merge(Episodes_df1, Episodes_df2, how='left', on="EpisodeID")

Episodes_df

# get month information from AdmitDate
Admitmonth_list = []
for date in Episodes_df["AdmitDate"]:
    Admitmonth_list.append(int(date.split('/')[0]))
    
# get month information from DischargeDate
Dischargemonth_list = []
for date in Episodes_df["DischargeDate"]:
    if str(date) == "nan":
        Dischargemonth_list.append(None)
    elif str(date[-1]) != "9":
        Dischargemonth_list.append(12)
    else:
        Dischargemonth_list.append(int(str(date).split('/')[0]))
        
# deal with none in dischargedate     
new_column = []
for i in range(len(Dischargemonth_list)):
    if Dischargemonth_list[i] == None:
        new_column.append('#')
        continue
    new_column.append((Admitmonth_list[i],Dischargemonth_list[i]+1))
Episodes_df['month_range'] = new_column

#create empty dict and keep the infomation into dictionay for further operation
_dict = dict()
for i in range(len(Episodes_df["PID"])):
    if Episodes_df['month_range'][i] == '#':
        continue
    if Episodes_df["PID"][i] not in _dict:
        _dict[Episodes_df["PID"][i]] = list(range(Episodes_df['month_range'][i][0],Episodes_df['month_range'][i][1]))
    else:
        _dict[Episodes_df["PID"][i]].extend(list(range(Episodes_df['month_range'][i][0],Episodes_df['month_range'][i][1])))

# sort pid order from small to large
_dict = sorted(_dict.items(), key = lambda x:x[0])
len(_dict)

# Create a two-dimensional array for keeping information in next operation
data=np.random.randn(1231,13)
data

# create a dataframe based on a two-dimensional array
df_ = pd.DataFrame(data,columns=['PID', 'M1', 'M2', 'M3','M4','M5','M6','M7','M8','M9','M10','M11','M12'],index=list(range(0,1231)),dtype=np.int64)
df_

# calculate the number of patient in each snf every month
# and keep the information into the dataframe we created
for i in range(len(_dict)):
    df_["PID"][i] = int(_dict[i][0])
    df_['M1'][i] = int(_dict[i][1].count(1))
    df_['M2'][i] = int(_dict[i][1].count(2))
    df_['M3'][i] = int(_dict[i][1].count(3))
    df_['M4'][i] = int(_dict[i][1].count(4))
    df_['M5'][i] = int(_dict[i][1].count(5))
    df_['M6'][i] = int(_dict[i][1].count(6))
    df_['M7'][i] = int(_dict[i][1].count(7))
    df_['M8'][i] = int(_dict[i][1].count(8))
    df_['M9'][i] = int(_dict[i][1].count(9))
    df_['M10'][i] = int(_dict[i][1].count(10))
    df_['M11'][i] = int(_dict[i][1].count(11))
    df_['M12'][i] = int(_dict[i][1].count(12))
df_

# load dataset including county information and save it into dataframe
county_complete = pd.read_csv('Data/county_complete.csv')  
county_complete

# combine two dataset
county_complete = county_complete[["ProviderName","PID","lat","lng","county_name"]]

# combine two dataset to attain a final dataset
result = pd.merge(df_, county_complete, how='left', on='PID')

result.to_csv("Data/curr_complete_dataset_delete_columns.csv",index=False)
