==> 09-Get-Yearly-Patient-Records.ipynb <==
# load library
import math
import numpy as np
import pandas as pd

# load data
data = pd.read_csv('Data/episodes_dur.csv')

data.head()

data = data[["PID","EpisodeID","ContractType","duration"]]

#calculate the number of patient in each snf in 2019
data_snf_patientnum_year = data.groupby(['PID'])[['EpisodeID']].count()

data_snf_patientnum_year

#load dataset
complete_dataset = pd.read_csv("Data/curr_complete_dataset_delete_columns.csv")

complete_dataset.head()

# combine the number of patient column into complete_dataset and create a new data frame
result = pd.merge(data_snf_patientnum_year, complete_dataset, how='left', on='PID')

result.rename(columns={'EpisodeID':'patient_number_year'}, inplace = True)

result.head()

# load CCs.csv dataset
cc_number = pd.read_csv("Data/CCs.csv")

# add relative columns from CCs.csv into current complete dataset
result2 = pd.merge(result, cc_number, how='left', on='PID')

result2.head()

result2.to_csv("Data/curr_complete_dataset_delete_columns.csv",index=False)
